requirements.txt
# Core InstrMCP Dependencies
fastmcp>=0.1.0
mcp>=1.0.0
pyyaml>=6.0
httpx>=0.25.0
nest-asyncio>=1.5.0

# QCodes Dependencies  
qcodes>=0.45.0
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0